# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "École Platform API",
    "description": "EdTech SaaS platform API for K-12 schools in Morocco.\n\n## Domains\n- **IAM** — Authentication, authorization, invitations, recovery, 2FA\n- **ERP** — Classes, enrollments, teacher assignments, attendance\n- **LMS** — Courses, assignments, submissions, results, content, activities, assessments\n- **Billing** — Invoices and payments\n- **COM** — Notifications, consents, parent feed\n- **AI** — Writing assistance, recommendations, KPIs\n\n## Authentication\nAll protected endpoints require a `Bearer` access token in the `Authorization` header.\nTokens are obtained via `POST /auth/login`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assessments": {
      "post": {
        "tags": [
          "lms-assessments"
        ],
        "summary": "Create an assessment",
        "description": "Create a formal assessment for a class.\n\nValidates:\n1. Class exists and is in the same school\n2. TCH: must be assigned to the class (ABAC)",
        "operationId": "create_assessment_api_v1_assessments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created assessment record",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "lms-assessments"
        ],
        "summary": "List assessments",
        "description": "List assessments. Optionally filtered by class_id and status.",
        "operationId": "list_assessments_api_v1_assessments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "class_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of assessments",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessments/{assessment_id}/publish": {
      "post": {
        "tags": [
          "lms-assessments"
        ],
        "summary": "Publish an assessment",
        "description": "Publish an assessment (change status from draft to published).",
        "operationId": "publish_assessment_api_v1_assessments__assessment_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assessment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published assessment",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessments/{assessment_id}/results": {
      "post": {
        "tags": [
          "lms-assessments"
        ],
        "summary": "Submit assessment result",
        "description": "Submit a result for an assessment (student self-report or auto-graded).\n\nValidates:\n1. Assessment exists and is published\n2. No duplicate result for same student+assessment (idempotent)",
        "operationId": "submit_assessment_result_api_v1_assessments__assessment_id__results_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assessment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentResultSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assessment result record",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentCreateRequest": {
        "properties": {
          "class_id": {
            "type": "string",
            "format": "uuid",
            "title": "Class Id"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "window_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "total_points": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Points",
            "default": 0
          },
          "status": {
            "type": "string",
            "pattern": "^(draft|published|closed)$",
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "class_id",
          "title"
        ],
        "title": "AssessmentCreateRequest"
      },
      "AssessmentResultSubmitRequest": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assessment Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "assessment_id"
        ],
        "title": "AssessmentResultSubmitRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}